Move types to matcher
